requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pinecone-client>=2.2.4
nltk>=3.8.1
pandas>=2.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0
tiktoken>=0.5.0
flask>=2.3.0 